tests/requirements-test.txt
pytest==7.4.0
pytest-cov==4.1.0
pytest-env==1.0.1
pytest-mock==3.11.1
pytest-xdist==3.3.1
coverage==7.3.0
requests-mock==1.11.0
freezegun==1.2.2
factory-boy==3.3.0
faker==19.3.1 